pyproject.toml
[tool.poetry]
name = "azure-pdf-parser"
version = "0.3.1"
description = ""
authors = ["CPR-tech-team <tech@climatepolicyradar.org>"]
packages = [{ include = "azure_pdf_parser", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
azure-ai-formrecognizer = "^3.2.1"
cpr-data-access = { git = "https://github.com/climatepolicyradar/data-access.git", tag = "0.4.0" }
requests = "^2.31.0"
langdetect = "^1.0.9"
pypdf = "^3.15.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.0.247"
black = "^22.1.0"
pyright = "^1.1.267"
pre-commit = "^2.17.0"
python-dotenv = "^0.19.2"
pytest = "^7.0.1"
pytest-dotenv = "^0.5.2"
nbmake = "^1.4.1"
httpx = "^0.27.0"
mock = "^5.1.0"


[tool.poetry.group.cli.dependencies]
click = "^8.1.7"

[tool.poetry.group.scripts.dependencies]
rich = "^13.7.1"
boto3 = "^1.34.111"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-p no:cacheprovider"
env_files = """
    .env.test
    .env
"""
markers = ["search", "unit"]


[tool.black]
line-length = 88
target-version = ["py39"]

[tool.ruff]
select = ["E", "F", "D"]
# Docstring Ignores:
# D100 - Missing docstring in public module
# D103 - Missing docstring in public function
# D104 - Missing docstring in public package
# D107 - Missing docstring in __init__
# D202 - No blank lines allowed after function docstring
# D203 - 1 blank line required before class docstring
# D213 - Multi-line docstring summary should start at the first line
# D400 - First line should end with a period
# D401 - First line should be in imperative mood
# D406 - Section name should end with a newline
# D407 - Missing dashed underline after section
# D413 - Missing blank line after last section
# D415 - First line should end with a period, question mark, or exclamation point
ignore = [
  "D100",
  "D103",
  "D104",
  "D107",
  "D202",
  "D203",
  "D212",
  "D400",
  "D401",
  "D406",
  "D407",
  "D413",
  "D415",
]
line-length = 88

# Ignore `E402` (import violations) in all `__init__.py` files, and in `path/to/file.py`.
[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["E501"]

[tool.pyright]
include = ["src"]
exclude = ["**/__pycache__"]
pythonVersion = "3.9"
